Rename bioacoustic_qc quality_control
