+++ src/senselab/audio/tasks/quality_control/__init__.py
@@ -1,17 +1,17 @@
 """Provides the API for bioacoustic quality control."""
 
-from .bioacoustic_qc import (
-    activity_to_dataset_taxonomy_subtree,
-    activity_to_evaluations,
-    activity_to_taxonomy_tree_path,
-    check_quality,
-    subtree_to_evaluations,
-)
 from .evaluate import (
     evaluate_audio,
     evaluate_batch,
     evaluate_dataset,
     evaluate_metric,
+)
+from .quality_control import (
+    activity_to_dataset_taxonomy_subtree,
+    activity_to_evaluations,
+    activity_to_taxonomy_tree_path,
+    check_quality,
+    subtree_to_evaluations,
 )
 
 __all__ = [
